pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geostats"
version = "0.3.0"
description = "A Python library for geostatistics and spatial data analysis"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Kyle Jones", email = "kyletjones@gmail.com"}
]
keywords = ["geostatistics", "kriging", "variogram", "spatial-statistics", "interpolation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "joblib>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
all = [
    "rasterio>=1.2.0",
    "netCDF4>=1.5.0",
    "geopandas>=0.10.0",
    "shapely>=1.8.0",
    "fiona>=1.8.0",
    "openpyxl>=3.0.0",
    "xgboost>=1.5.0",
    "plotly>=5.0.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "pydantic>=1.10.0",
    "python-multipart>=0.0.5",
    "jinja2>=3.0.0",
    "click>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/kylejones200/geostats"
Documentation = "https://geostats.readthedocs.io"
Repository = "https://github.com/kylejones200/geostats"

[project.scripts]
geostats-run = "geostats.cli:run"
geostats-validate = "geostats.cli:validate"
geostats-init = "geostats.cli:init"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=geostats --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E203", "E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
